# openapi/api.yml
openapi: 3.0.0
info:
  title: GigaChill
  version: 0.1.0
servers:
  - url: https://giga-chill.ru/api
  - url: http://localhost:8081
security:
  - CookieAuth: []
...
paths:
...
  /events/{eventId}/tasks:
    get:
      summary: Получить задачи
      tags: [Tasks]
      operationId: getTasks
      parameters:
        - $ref: "#/components/parameters/eventId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tasks"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    post:
      summary: Создать задачу
      tags: [Tasks]
      operationId: createTask
      parameters:
        - $ref: "#/components/parameters/eventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/tasks/{taskId}:
    get:
      summary: Получить задачу
      tags: [Tasks]
      operationId: getTask
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskWithShoppingLists"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    patch:
      summary: Обновить задачу
      tags: [Tasks]
      operationId: updateTask
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      summary: Удалить задачу
      tags: [Tasks]
      operationId: deleteTask
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/tasks/{taskId}/executor:
    put:
      summary: Указать исполнителя
      description: Поле executor_id может быть null - это означает,
        что у задачи нет исполнителя. В этом случае задача переходит в статус "открыта".
      tags: [Tasks]
      operationId: setTaskExecutor
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSetExecutor"
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/InvalidDataError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
  /events/{eventId}/tasks/{taskId}/take-in-work:
    post:
      summary: Взять задачу в работу
      description: Если исполнитель не указан, то тот кто взял задачу в работу становится её исполнителем.
      tags: [Tasks]
      operationId: takeTaskInWork
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /events/{eventId}/tasks/{taskId}/send-for-review:
    post:
      summary: Отправить задачу на проверку
      tags: [Tasks]
      operationId: sendTaskForReview
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskSendForReviewRequest"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /events/{eventId}/tasks/{taskId}/review:
    post:
      summary: Подтвердить или отклонить задачу
      tags: [Tasks]
      operationId: reviewTask
      parameters:
        - $ref: "#/components/parameters/eventId"
        - $ref: "#/components/parameters/taskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskReviewRequest"
      responses:
        "204":
          description: No Content
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
...
components:
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
...